Add rename LoadingApp -> LoadingSpinner and remove use of makeStyles
