--- packages/common/src/ui/components/LoadingApp.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import { CircularProgress, Typography } from '@material-ui/core';
-import makeStyles from '@material-ui/styles/makeStyles';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: 'column',
-    position: 'fixed',
-    top: 0,
-    left: 0,
-    width: '100%',
-    height: '100%',
-  },
-  text: {
-    marginTop: theme?.spacing?.(3),
-  },
-}));
-
-export const LoadingApp: React.FC = () => {
-  const classes = useStyles();
-  return (
-    <div className={classes.root}>
-      <CircularProgress />
-      <Typography className={classes.text}>Loading...</Typography>
-    </div>
-  );
-};
