+++ packages/common/src/ui/components/LoadingSpinner.tsx
@@ -1,0 +1,30 @@
+import React, { FC } from 'react';
+import { CircularProgress, Typography } from '@material-ui/core';
+import { Box, styled } from '@material-ui/system';
+
+const Container = styled(Box)({
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  flexDirection: 'column',
+  position: 'fixed',
+  top: 0,
+  left: 0,
+  width: '100%',
+  height: '100%',
+});
+
+const StyledText = styled(Typography)(({ theme }) => ({
+  text: {
+    marginTop: theme.spacing(3),
+  },
+}));
+
+export const LoadingSpinner: FC = () => {
+  return (
+    <Container>
+      <CircularProgress />
+      <StyledText>Loading...</StyledText>
+    </Container>
+  );
+};
